requirements.txt
requests
Faker
matplotlib
python-dotenv